requirements.txt
tqdm 
python-vlc
bs4
selenium